Fix ProfilesController#update does not handle errors on the update
It seems the current implementation expects the `profile` variable
may be changed into nil when the update failed.  But it must be always
truthy and it raises an exception on failed, instead.

It's better to use begin-rescue instead of if-else to handle exception
and emit an error to the users.

This error was found via `Ruby::UnreachableBranch` error from Steep :-)

Note: The `UnreachableBranch` is not emmited by default.  Locally, I
enabled the "strict" settings of Steep to find this bug.